remove warning
git-svn-id: http://svn.research.cc.gatech.edu/kaos/ffs/trunk@7386 8ba6a8b9-845b-0410-94f0-974be2f5df47

+++ fm/fm_formats.c
@@ -3303,7 +3303,7 @@ dump_subfield(void*base, FMFormat f, dstate s, int data_offset, void* parent_bas
 #if defined (__INTEL_COMPILER)
 #  pragma warning (disable: 810)
 #endif
-	    new_offset = ptr_value;
+	  new_offset = (long) ptr_value;
 #if defined (__INTEL_COMPILER)
 #  pragma warning (default: 810)
 #endif
